AppData file
AppData is a data format used in GNOME and KDE software installers to
add the application description and some screenshots.

See the specification for more details:

  http://people.freedesktop.org/~hughsient/appdata/

Install the file submitted to us by hughsie which I'd like to thank.
Later when intltool will be integrated with CMake we'll be able to
extract strings for translation and let our translators do the rest.

Issue: 8